move nodes and stats out of explorer
