+++ packages/playground/src/components/dashboard/user_farms.vue
@@ -119,7 +119,7 @@ import { StrKey } from "stellar-sdk";
 import { onMounted, ref } from "vue";
 
 import { gridProxyClient } from "@/clients";
-import CardDetails from "@/explorer/components/node_details_cards/card_details.vue";
+import CardDetails from "@/components/node_details_cards/card_details.vue";
 import { useGrid, useProfileManager } from "@/stores";
 import { createCustomToast, ToastType } from "@/utils/custom_toast";
 import {
