+++ packages/playground/src/components/node_details_cards/gpu_details_card.vue
@@ -31,7 +31,7 @@ import type { GPUCard, GridNode } from "@threefold/gridproxy_client";
 import { onMounted, type PropType, ref } from "vue";
 import { watch } from "vue";
 
-import type { NodeDetailsCard } from "@/explorer/utils/types";
+import type { NodeDetailsCard } from "@/types";
 import { createCustomToast, ToastType } from "@/utils/custom_toast";
 
 import CardDetails from "./card_details.vue";
