+++ packages/playground/src/components/node_details_cards/location_details_card.vue
@@ -14,7 +14,7 @@
 import type { GridNode } from "@threefold/gridproxy_client";
 import { type PropType, ref } from "vue";
 
-import type { NodeDetailsCard } from "@/explorer/utils/types";
+import type { NodeDetailsCard } from "@/types";
 
 import CardDetails from "./card_details.vue";
