+++ packages/playground/src/components/node_details_cards/twin_details_card.vue
@@ -1,18 +1,18 @@
 <template>
-  <card-details :loading="loading" title="Farm Details" :items="farmFields" icon="mdi-webpack" />
+  <card-details :loading="loading" title="Node Twin Details" :items="twinFields" icon="mdi-account" />
 </template>
 
 <script lang="ts">
 import type { GridNode } from "@threefold/gridproxy_client";
 import { onMounted, type PropType, ref } from "vue";
 
-import type { NodeDetailsCard } from "@/explorer/utils/types";
+import type { NodeDetailsCard } from "@/types";
 import { createCustomToast, ToastType } from "@/utils/custom_toast";
 
 import CardDetails from "./card_details.vue";
 
 export default {
-  name: "FarmDetailsCard",
+  name: "TwinDetailsCard",
   components: { CardDetails },
   props: {
     node: {
@@ -23,11 +23,11 @@ export default {
 
   setup(props) {
     const loading = ref<boolean>(false);
-    const farmFields = ref<NodeDetailsCard[]>();
+    const twinFields = ref<NodeDetailsCard[]>();
 
     const mount = () => {
       loading.value = true;
-      farmFields.value = getFarmDetails();
+      twinFields.value = getNodeTwinDetailsCard();
       loading.value = false;
     };
 
@@ -38,22 +38,22 @@ export default {
       createCustomToast("Copied!", ToastType.success);
     };
 
-    const getFarmDetails = (): NodeDetailsCard[] => {
+    const getNodeTwinDetailsCard = (): NodeDetailsCard[] => {
       return [
-        { name: "ID", value: props.node.farmId.toString() },
-        { name: "Name", value: props.node.farm.name },
+        { name: "ID", value: props.node.twin.twinId.toString() },
         {
-          name: "Stellar Address",
-          value: props.node.farm.stellarAddress,
+          name: "Account ID",
+          value: props.node.twin.accountId,
           icon: "mdi-content-copy",
           callback: copy,
-          hint: "Copy the stellar address to the clipboard.",
+          hint: "Copy the account id to the clipboard.",
         },
+        { name: "Relay", value: props.node.twin.relay },
       ];
     };
 
     return {
-      farmFields,
+      twinFields,
       loading,
     };
   },
