+++ packages/playground/src/utils/validators.ts
@@ -4,7 +4,7 @@ import type { IsEmailOptions } from "validator/lib/isEmail";
 import type { IsFQDNOptions } from "validator/lib/isFQDN";
 import type { IsURLOptions } from "validator/lib/isURL";
 
-import { toBytes } from "@/explorer/utils/helpers";
+import { toBytes } from "@/utils/get_nodes";
 
 export function required(msg: string) {
   return (value: string) => {
