+++ packages/playground/src/views/stats.vue
